Add files via upload
Housing design for Iris 7. 

Description.

It is possible to assemble the IRIS with the included kit and middle layer alone. 

The top layer is a design element that allows you to combine the design with the palmrest, but you may choose to install and combine them for completeness. 

The palmrest layer is intended to be joined to the top layer by the palmrest frame. 

Parts Required

In addition to the screws provided by IRIS, the following materials are required for assembly 

- 20EA of M2 16mm, or 15mm screws
- 4 EA of 9 mm diameter, 3 mm high magnets
- 24EA 3mm diameter, 2mm high magnets
- 4 EA of 3.2 mm diameter, 3 mm high M2 brass brackets
- 30 cm of fabric cable ties 
- (maybe) a file, screwdriver, and pliers. 

Translated with www.DeepL.com/Translator (free version)